+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 
+## [in progress]
+
+### Deprecated or removed
+
+- Parameter `thread_pool_executor` to `BiglistBase.__init__` and `ParquetBiglist.new`.
+
+
 ## [0.7.3] - 2022-12-27
 
 - Upgrade dependency `upathlib`, removing formats 'json-z' and 'json-zstd'.
@@ -12,7 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [0.7.2] - 2022-12-14
 
-## Enhanced
+### Enhanced
 
 - Enhancement and refinement of type annotations.
 - Refine the "generic type" annotations for the classes.